Update tests to the new protocol model

+++ src/bc/model.rs
@@ -80,7 +80,7 @@ pub enum EncryptionProtocol {
 
 #[derive(Debug)]
 pub struct BcContext {
-    pub(super) in_bin_mode: HashSet<u32>,
+    pub(super) in_bin_mode: HashSet<u16>,
     // Arc<Mutex<EncryptionProtocol>> because it is shared between context
     // and connection for deserialisation and serialistion respectivly
     pub(super) encryption_protocol: Arc<Mutex<EncryptionProtocol>>,
